# BitMapFilter altered to use ref int