This PR changes the `BitMapFilter` to use ref int. This simplifies the code a bit and allows to make it less `unsafe`. Also, by basing all calls on `ref int`, should allow to use `Interlocked.Or` and make the filter concurrency/atomic friendly. An addiitonal side effect is faster `MayContainAny`

### Benchmarks

#### Before

| Method                             | Mean          | Error       | StdDev      | Code Size | Allocated |
|----------------------------------- |--------------:|------------:|------------:|----------:|----------:|
| Or_BitMapFilter_Of1                |    38.9126 ns |   0.3472 ns |   0.3247 ns |     235 B |         - |
| Or_BitMapFilter_Of2                |    82.1053 ns |   1.6279 ns |   1.5228 ns |     459 B |         - |
| Or_BitMapFilter_OfN_128            | 9,127.0372 ns | 116.5773 ns | 103.3428 ns |     332 B |         - |
| MayContainAny_BitMapFilter_OfN_128 |     0.9688 ns |   0.0146 ns |   0.0137 ns |     435 B |         - |

#### After

| Method                             | Mean          | Error       | StdDev      | Code Size | Allocated |
|----------------------------------- |--------------:|------------:|------------:|----------:|----------:|
| Or_BitMapFilter_Of1                |    38.5533 ns |   0.4279 ns |   0.4003 ns |     235 B |         - |
| Or_BitMapFilter_Of2                |    80.3172 ns |   1.3885 ns |   1.2309 ns |     459 B |         - |
| Or_BitMapFilter_OfN_128            | 9,032.1915 ns | 153.9109 ns | 143.9684 ns |     332 B |         - |
| MayContainAny_BitMapFilter_OfN_128 |     0.6839 ns |   0.0129 ns |   0.0121 ns |     332 B |         - |

### Diff for MayContainAny

![image](https://github.com/NethermindEth/Paprika/assets/519707/a0d0e1df-3cfb-4305-8947-07373ca34116)